dlproject.yaml: set build_type and arch for tools' build_requirements
When building tools, the build_requirements of the tools don't have to be the
same architecture, for instance, it's possible to use a 64-bit Release bison to
build a 32-bit Debug Doxygen.

Do this with per-package settings, because the recipes in conan-center-index
don't really support split host/build profiles yet.

This fixes problems like not being able to make a 32-bit Doxygen on Windows.