Jenkinsfile: Use 'def' to make local variables
By default, in a script, referencing a variable will cause it to become part of
the script's "binding", which essentially makes it a global. Some of the
variables in the scripting stages were mixing across parallel runs, causing
mixups in the run, crossing the names of test output files and the like.

See: https://stackoverflow.com/a/60734868/11996393
See: https://stackoverflow.com/a/185879/11996393