cmake: Can also now build from sources
CMake can be built from sources on demand, or automatically when
settings.os is not Linux, Macos, or Windows.

The code was brought over from the 3.x.x recipe, leaving out the support
for bootstrapping. One can manually build or obtain a CMake to build
this recipe with.

The recipe name can be changed with CMAKE_RECIPE_NAME in order to keep a
CMake Conan package with a different name, to be used in tool_requires
when building newer CMake versions. Having a bootstrapping strategy like
that removes the require to keep or maintain the bootstrap code.