tool_require the CMake tool on Linux, macOS and Windows
To alleviate problems with CMake being available on various builders,
just always include it when building anything but CMake.

The CMake tool is a binary download on Linux, macOS, and Windows, so
there are no requirements for building it. Even if it's not present,
it's fast to download and unpack.

But...use a separate config for the builds/uploads of CMake itself. If
CMake is in the requirements for building a Conan package of itself to
upload, the tool_requires of CMake might be found instead of the newly
built CMake, causing the test_package to fail.