Add back the _write_project_config method to the b2 recipe
When CI ran after a merge conflict between the upstream conan-io
b2 recipe and the DL conan-center-index, AIX and Sparc platforms
were failing to build b2. After looking through logs, it seemed like
there was an error because the b2 recipe didn't have a _write_project_config()
method defined. The upstream recipe did have the function defined, and we
removed it in this commit:
https://octocat.dlogics.com/datalogics/conan-center-index/commit/d8b23aeb2eeeb7a1f2191ff8ac247214ce59a7a4.
Let's try adding it back to see if it helps to get CI passing.